# improve cell read performance by optimizing xml parsing