# PR Details

improve cell read performance by optimizing  unmarshal for xlsxC struct

## Description

<!--- Describe your changes in detail -->


## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

Improve the cell reading performance of large documents

## How Has This Been Tested

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- See how your change affects other areas of the code, etc. -->

Performance testing was conducted on an AMD64 WSL using Go's benchmark tool. The benchmarks were run BenchmarkRows 10 times each to ensure statistical significance. The following results were observed:
goos: linux
goarch: amd64
pkg: github.com/xuri/excelize/v2
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
```
        │ bashline_1.txt │              new_1.txt              │
        │     sec/op     │   sec/op     vs base                │
Rows-12      164.7µ ± 5%   131.1µ ± 2%  -20.41% (p=0.000 n=10)

        │ bashline_1.txt │              new_1.txt              │
        │      B/op      │     B/op      vs base               │
Rows-12     72.20Ki ± 0%   66.44Ki ± 0%  -7.99% (p=0.000 n=10)

        │ bashline_1.txt │              new_1.txt              │
        │   allocs/op    │  allocs/op   vs base                │
Rows-12      1.903k ± 0%   1.658k ± 0%  -12.87% (p=0.000 n=10)
```

Due to the small size of Book1.xlsx, I replaced it with TestStreamWriter.xlsx(Get after run TestStreamWriter) and re-executed it:
```
       │ bashline_2.txt │             new_2.txt              │
        │     sec/op     │   sec/op    vs base                │
Rows-12       5.662 ± 3%   3.970 ± 3%  -29.88% (p=0.000 n=10)

        │ bashline_2.txt │              new_2.txt               │
        │      B/op      │     B/op      vs base                │
Rows-12     2.387Gi ± 0%   2.044Gi ± 0%  -14.38% (p=0.000 n=10)

        │ bashline_2.txt │              new_2.txt              │
        │   allocs/op    │  allocs/op   vs base                │
Rows-12      64.82M ± 0%   52.02M ± 0%  -19.75% (p=0.000 n=10)
```
## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.